removed ODEDynamics and resolved circular imports
