--- chirho/dynamical/handlers/ODE/__init__.py
@@ -1,2 +1,0 @@
-from . import solvers  # noqa F401
-from .ode import ODESolver  # noqa F401
