--- chirho/dynamical/handlers/ODE/ode.py
@@ -1,5 +1,0 @@
-from chirho.dynamical.handlers.solver import Solver
-
-
-class ODESolver(Solver):
-    pass
