--- chirho/dynamical/handlers/ODE/solvers/__init__.py
@@ -1,1 +1,0 @@
-from .torchdiffeq import TorchDiffEq  # noqa: F401
