--- chirho/dynamical/handlers/ODE/solvers/torchdiffeq.py
@@ -1,16 +1,0 @@
-from chirho.dynamical.handlers.ODE.ode import ODESolver
-
-
-class TorchDiffEq(ODESolver):
-    def __init__(self, rtol=1e-7, atol=1e-9, method=None, options=None):
-        self.rtol = rtol
-        self.atol = atol
-        self.method = method
-        self.options = options
-        self.odeint_kwargs = {
-            "rtol": rtol,
-            "atol": atol,
-            "method": method,
-            "options": options,
-        }
-        super().__init__()
