+++ chirho/dynamical/handlers/solver.py
@@ -1,5 +1,3 @@
-from __future__ import annotations
-
 import pyro
 
 
@@ -7,3 +5,18 @@ class Solver(pyro.poutine.messenger.Messenger):
     def _pyro_simulate(self, msg) -> None:
         # Overwrite the solver in the message with the enclosing solver when used as a context manager.
         msg["kwargs"]["solver"] = self
+
+
+class TorchDiffEq(Solver):
+    def __init__(self, rtol=1e-7, atol=1e-9, method=None, options=None):
+        self.rtol = rtol
+        self.atol = atol
+        self.method = method
+        self.options = options
+        self.odeint_kwargs = {
+            "rtol": rtol,
+            "atol": atol,
+            "method": method,
+            "options": options,
+        }
+        super().__init__()
