+++ chirho/dynamical/internals/__init__.py
@@ -1,5 +1,5 @@
-from . import ODE  # noqa: F401
+# Include only imports that are needed for registering dispatches.
+
 from . import dynamical  # noqa: F401
 from . import indexed  # noqa: F401
-from . import interruption  # noqa: F401
-from . import interventional  # noqa: F401
+from . import solver  # noqa: F401
