+++ chirho/dynamical/internals/dynamical.py
@@ -1,10 +1,15 @@
+from __future__ import annotations
+
 import functools
-from typing import Optional, TypeVar
+from typing import TYPE_CHECKING, TypeVar
 
 import torch
 
-from chirho.dynamical.handlers.solver import Solver
 from chirho.dynamical.ops.dynamical import Dynamics, State, Trajectory
+
+if TYPE_CHECKING:
+    from chirho.dynamical.handlers.solver import Solver
+
 
 S = TypeVar("S")
 T = TypeVar("T")
@@ -12,26 +17,17 @@ T = TypeVar("T")
 
 @functools.singledispatch
 def simulate_trajectory(
+    solver: "Solver",  # Quoted type necessary w/ TYPE_CHECKING to avoid circular import error
     dynamics: Dynamics[S, T],
     initial_state: State[T],
     timespan: T,
-    *,
-    solver: Optional[Solver] = None,
     **kwargs,
 ) -> Trajectory[T]:
     """
     Simulate a dynamical system.
     """
-    if solver is None:
-        raise ValueError(
-            "`simulate_trajectory` requires a solver. To specify a solver, use the keyword argument `solver` in"
-            " the call to `simulate_trajectory` or use with a solver effect handler as a context manager. "
-            "For example,"
-            "\n \n `with TorchDiffEq():` \n"
-            "\t `simulate_trajectory(dynamics, initial_state, start_time, end_time)`"
-        )
     raise NotImplementedError(
-        f"simulate_trajectory not implemented for type {type(dynamics)}"
+        f"simulate_trajectory not implemented for solver of type {type(solver)}"
     )
