+++ chirho/dynamical/internals/interruption.py
@@ -1,15 +1,20 @@
+from __future__ import annotations
+
 import functools
-from typing import List, Optional, Tuple, TypeVar
+from typing import TYPE_CHECKING, List, Optional, Tuple, TypeVar
 
 import pyro
 
-from chirho.dynamical.handlers.interruption import (
+from chirho.dynamical.handlers.interruption.interruption import (
     DynamicInterruption,
     Interruption,
     StaticInterruption,
 )
-from chirho.dynamical.handlers.solver import Solver
 from chirho.dynamical.ops.dynamical import Dynamics, State, simulate
+
+if TYPE_CHECKING:
+    from chirho.dynamical.handlers.solver import Solver
+
 
 S = TypeVar("S")
 T = TypeVar("T")
@@ -18,16 +23,16 @@ T = TypeVar("T")
 # Separating out the effectful operation from the non-effectful dispatch on the default implementation
 @pyro.poutine.runtime.effectful(type="simulate_to_interruption")
 def simulate_to_interruption(
+    solver: "Solver",  # Quoted type necessary w/ TYPE_CHECKING to avoid circular import error
     dynamics: Dynamics[S, T],
     start_state: State[T],
     start_time: T,
     end_time: T,
     *,
-    solver: Optional[Solver] = None,
-    next_static_interruption: Optional["StaticInterruption"] = None,
-    dynamic_interruptions: List["DynamicInterruption"] = [],
+    next_static_interruption: Optional[StaticInterruption] = None,
+    dynamic_interruptions: List[DynamicInterruption] = [],
     **kwargs,
-) -> Tuple[State[T], Tuple["Interruption", ...], T]:
+) -> Tuple[State[T], Tuple[Interruption, ...], T]:
     """
     Simulate a dynamical system until the next interruption. This will be either one of the passed
     dynamic interruptions, the next static interruption, or the end time, whichever comes first.
@@ -36,11 +41,11 @@ def simulate_to_interruption(
     """
 
     interruptions, interruption_time = get_next_interruptions(
+        solver,
         dynamics,
         start_state,
         start_time,
         end_time,
-        solver=solver,
         next_static_interruption=next_static_interruption,
         dynamic_interruptions=dynamic_interruptions,
         **kwargs,
@@ -56,16 +61,16 @@ def simulate_to_interruption(
 
 
 def get_next_interruptions(
+    solver: "Solver",  # Quoted type necessary w/ TYPE_CHECKING to avoid circular import error
     dynamics: Dynamics[S, T],
     start_state: State[T],
     start_time: T,
     end_time: T,
     *,
-    solver: Optional[Solver] = None,
-    next_static_interruption: Optional["StaticInterruption"] = None,
-    dynamic_interruptions: List["DynamicInterruption"] = [],
+    next_static_interruption: Optional[StaticInterruption] = None,
+    dynamic_interruptions: List[DynamicInterruption] = [],
     **kwargs,
-) -> Tuple[Tuple["Interruption", ...], T]:
+) -> Tuple[Tuple[Interruption, ...], T]:
     nodyn = len(dynamic_interruptions) == 0
     nostat = next_static_interruption is None
 
@@ -84,10 +89,10 @@ def get_next_interruptions(
         return (next_static_interruption,), next_static_interruption.time  # type: ignore
     else:
         return get_next_interruptions_dynamic(  # type: ignore
+            solver,  # type: ignore
             dynamics,  # type: ignore
             start_state,  # type: ignore
             start_time,  # type: ignore
-            solver=solver,
             next_static_interruption=next_static_interruption,
             dynamic_interruptions=dynamic_interruptions,
             **kwargs,
@@ -99,14 +104,12 @@ def get_next_interruptions(
 # noinspection PyUnusedLocal
 @functools.singledispatch
 def get_next_interruptions_dynamic(
+    solver: "Solver",  # Quoted type necessary w/ TYPE_CHECKING to avoid circular import error
     dynamics: Dynamics[S, T],
     start_state: State[T],
     start_time: T,
     next_static_interruption: StaticInterruption,
     dynamic_interruptions: List[DynamicInterruption],
-    *,
-    solver: Optional[Solver] = None,
-    **kwargs,
 ) -> Tuple[Tuple[Interruption, ...], T]:
     raise NotImplementedError(
         f"get_next_interruptions_dynamic not implemented for type {type(dynamics)}"
