+++ chirho/dynamical/internals/interventional.py
@@ -1,5 +1,3 @@
-from __future__ import annotations
-
 from typing import TypeVar
 
 from chirho.dynamical.ops.dynamical import State
