+++ chirho/dynamical/internals/solver/__init__.py
