+++ chirho/dynamical/ops/__init__.py
@@ -1,2 +1,1 @@
-from . import ODE  # noqa: F401
 from .dynamical import Dynamics, State, Trajectory, simulate  # noqa: F401
