+++ tests/dynamical/dynamical_fixtures.py
@@ -4,13 +4,12 @@ import pyro
 import torch
 from pyro.distributions import Normal, Uniform, constraints
 
-from chirho.dynamical.ops.dynamical import State, Trajectory
-from chirho.dynamical.ops.ODE import ODEDynamics
+from chirho.dynamical.ops.dynamical import Dynamics, State, Trajectory
 
 T = TypeVar("T")
 
 
-class UnifiedFixtureDynamics(ODEDynamics):
+class UnifiedFixtureDynamics(Dynamics):
     def __init__(self, beta=None, gamma=None):
         super().__init__()
