+++ tests/dynamical/test_noop_interruptions.py
@@ -9,7 +9,7 @@ from chirho.dynamical.handlers import (
     StaticInterruption,
     StaticIntervention,
 )
-from chirho.dynamical.handlers.ODE.solvers import TorchDiffEq
+from chirho.dynamical.handlers.solver import TorchDiffEq
 from chirho.dynamical.ops import State, simulate
 
 from .dynamical_fixtures import UnifiedFixtureDynamics, check_states_match
