+++ tests/dynamical/test_solver.py
@@ -5,7 +5,7 @@ import pytest
 import torch
 
 from chirho.dynamical.handlers import SimulatorEventLoop
-from chirho.dynamical.handlers.ODE.solvers import TorchDiffEq
+from chirho.dynamical.handlers.solver import TorchDiffEq
 from chirho.dynamical.ops import State, simulate
 
 from .dynamical_fixtures import bayes_sir_model, check_states_match
